+++ Tests/Unit/EmogrifierTest.php
@@ -950,13 +950,30 @@ class EmogrifierTest extends \PHPUnit_Framework_TestCase
      *
      * @expectedException \InvalidArgumentException
      */
-    public function emogrifyForInvalidCssSelectorThrowsException()
+    public function emogrifyInDebugModeForInvalidCssSelectorThrowsException()
     {
         $this->subject->setHtml(
             '<html><style type="text/css">p{color:red;} <style data-x="1">html{cursor:text;}</style></html>'
         );
 
+        $this->subject->setDebug(true);
         $this->subject->emogrify();
+    }
+
+    /**
+     * @test
+     */
+    public function emogrifyNotInDebugModeIgnoresInvalidCssSelectors()
+    {
+        $html = '<html><style type="text/css">' .
+            'p{color:red;} <style data-x="1">html{cursor:text;} p{background-color:blue;}</style> ' .
+            '<body><p></p></body></html>';
+
+        $this->subject->setHtml($html);
+
+        $html = $this->subject->emogrify();
+        self::assertContains('color: red', $html);
+        self::assertContains('background-color: blue', $html);
     }
 
     /**