+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## x.y.z (unreleased)
 
 ### Added
+- Debug mode. Throw debug exceptions only if debug is active.
+  ([#392](https://github.com/jjriv/emogrifier/pull/392))
 
 
 ### Changed